# Conflicts:
#	views/template.php

+++ controllers/gestiontournoiController.class.php
@@ -41,6 +41,11 @@ class gestiontournoiController extends template{
 				$v->assign("tournoi",$tournament);
 				$v->assign("content", "Gérer votre tournoi");
 
+				if($tournament->getStatus()=="-1")
+					$v->assign("verrouillage"," disabled ");
+				else
+					$v->assign("verrouillage"," ");
+
 				/* MAJ effectuées auparavant */
 				if(isset($_SESSION['referer_method'])){
 					
@@ -115,7 +120,7 @@ class gestiontournoiController extends template{
 		$tournament = $tournamentBDD->getTournamentWithLink($link);
 
 		if(!!$link && is_bool(strpos($link, 'null')) && $tournament !== false 
-				&& $tournament->getIdUserCreator()==$this->connectedUser->getId()){
+				&& $tournament->getIdUserCreator()==$this->connectedUser->getId() && date('Ymd')<date('Ymd',$tournament->getStartDate())){
 
 		    $newtournament = new tournament($checkedDatas);
 
@@ -135,12 +140,12 @@ class gestiontournoiController extends template{
 			    $tournamentBDD->setTournament($tournament, $newtournament);
 
 				$_SESSION['referer_method']="update";
-			}
-			else
-				$_SESSION['referer_method']="error";
-
-			header("Location: ".$_SERVER['HTTP_REFERER']."");
+			}			
 		}
+		if(!isset($_SESSION['referer_method']))
+			$_SESSION['referer_method']="error";
+		
+		header("Location: ".$_SERVER['HTTP_REFERER']."");
 	}
 
 	//Methode présente dans Controller et non template car on ne peut faire de MAJ qu'ici
@@ -205,22 +210,43 @@ class gestiontournoiController extends template{
   	public function deleteTourAction(){
 
         $args = array(
-            'nom' => FILTER_SANITIZE_STRING
+            'link' => FILTER_SANITIZE_STRING
         );
 
         $filteredinputs = filter_input_array(INPUT_GET, $args);
 
         $tournamentBDD = new tournamentManager();
-        $tournoi = $tournamentBDD->getFilteredTournaments($filteredinputs['nom'])[0];
+        $tournoi = $tournamentBDD->getTournamentWithLink($filteredinputs['link']);
     
-        if($tournoi){
+        if($tournoi && $tournoi->getIdUserCreator() == $_id){
             $tournamentBDD->deleteTour($tournoi);
-        	
         }
         else
             return null;
 
     }
 
+    public function mailMember(){
+
+    	//FILTER_SANITIZE_STRING Remove all HTML tags from a string
+	    $args = array( 'message' => FILTER_SANITIZE_STRING, 
+	    			   'link' => FILTER_SANITIZE_STRING
+	    			);
+		$filteredinputs = filter_input_array(INPUT_POST, $args);
+
+   		$tournamentBDD = new tournamentManager();
+       	$tournoi = $tournamentBDD->getTournamentWithLink($filteredinputs['link']);
+
+		if($filteredinputs && $tournoi->getIdUserCreator() == $_id){
+
+			$registerBDD = new registerManager();
+			$listeuser = $registerBDD->getTournamentParticipants($tournoi);
+
+			foreach ($listeuser as $user) {
+				$this->envoiMail($user->getEmail(),'Un organisateur de tournoi vous a envoyé un message.',$filteredinputs['message']);
+			}
+ 
+    	}
+    }
 
 }
